move modern design gigfunding related react components into Gigfunding directory for easier location and reuse
