+++ client/app/components/elements/Gigfunding/Text/CenteredTextContent.js
